move Admin components out of Geral subfolder
